+++ .gitignore
@@ -35,3 +35,7 @@ out/
 
 ### VS Code ###
 .vscode/
+
+
+### profile
+src/main/resources/application.properties
