--- src/main/java/billage/BillageApplication.java
@@ -1,13 +1,0 @@
-package billage;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BillageApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BillageApplication.class, args);
-    }
-
-}
