--- src/main/java/billage/exception/ControllerAdvice.java
@@ -1,16 +1,0 @@
-package billage.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ControllerAdvice {
-
-    @ExceptionHandler(IllegalArgumentException.class)
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse handleIllegalArgumentException(RuntimeException e) {
-        return new ErrorResponse(e.getMessage());
-    }
-}
