--- src/main/java/billage/exception/ErrorResponse.java
@@ -1,14 +1,0 @@
-package billage.exception;
-
-public class ErrorResponse {
-
-    private final String message;
-
-    public ErrorResponse(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
