--- src/main/java/billage/repository/MemberRepository.java
@@ -1,23 +1,0 @@
-package billage.repository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-import org.springframework.stereotype.Repository;
-
-import billage.domain.Member;
-
-import lombok.RequiredArgsConstructor;
-
-@Repository
-@RequiredArgsConstructor
-public class MemberRepository {
-
-    @PersistenceContext
-    private final EntityManager em;
-
-    public Long save(Member member) {
-        em.persist(member);
-        return member.getId();
-    }
-}
