+++ src/main/java/sogorae/BillageApplication.java
@@ -1,0 +1,13 @@
+package sogorae;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BillageApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(BillageApplication.class, args);
+    }
+
+}
