+++ src/main/java/sogorae/billage/exception/ErrorResponse.java
@@ -1,0 +1,14 @@
+package sogorae.billage.exception;
+
+public class ErrorResponse {
+
+    private final String message;
+
+    public ErrorResponse(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
