+++ src/main/java/sogorae/billage/repository/MemberRepository.java
@@ -1,0 +1,23 @@
+package sogorae.billage.repository;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+import org.springframework.stereotype.Repository;
+
+import sogorae.billage.domain.Member;
+
+import lombok.RequiredArgsConstructor;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberRepository {
+
+    @PersistenceContext
+    private final EntityManager em;
+
+    public Long save(Member member) {
+        em.persist(member);
+        return member.getId();
+    }
+}
