--- src/test/java/billage/BillageApplicationTests.java
@@ -1,13 +1,0 @@
-package billage;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BillageApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
