--- src/test/java/billage/repository/MemberRepositoryTest.java
@@ -1,32 +1,0 @@
-package billage.repository;
-
-import static org.assertj.core.api.Assertions.*;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-import billage.domain.Member;
-
-@SpringBootTest
-@Transactional
-public class MemberRepositoryTest {
-
-    @Autowired
-    private MemberRepository memberRepository;
-
-    @Test
-    @DisplayName("회원 저장 기능을 검증한다.")
-    void save() {
-        // given
-        Member member = new Member("sojukang@gmail.com", "sojukang", "12345678");
-
-        // when
-        Long savedId = memberRepository.save(member);
-
-        // then
-        assertThat(savedId).isNotNull();
-    }
-}
