+++ src/test/java/sogorae/billage/BillageApplicationTests.java
@@ -1,0 +1,13 @@
+package sogorae.billage;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BillageApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
